update move method

+++ src/com/huan/dane/dota/BaseObject.java
@@ -3,27 +3,27 @@ package com.huan.dane.dota;
 import java.awt.image.BufferedImage;
 
 public abstract class BaseObject {
-    int x;
-    int y;
+    float x;
+    float y;
 
     int width;
     int height;
 
     BufferedImage image;
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
